brief_site/requirements.txt
asgiref==3.2.10
coverage==5.2.1
Django==3.1
django-nose==1.4.7
djangorestframework==3.11.1
feedparser==5.2.1
mysql-client==0.0.1
nose==1.3.7
PyMySQL==0.10.0
pytz==2020.1
sqlparse==0.3.1
